refactor: graphQL
