+++ server/src/graphql/index.ts
@@ -1,0 +1,7 @@
+import { typeDefs } from "./typeDefs"
+import { resolvers } from "./resolvers"
+
+export {
+    typeDefs,
+    resolvers
+}
