+++ server/src/graphql/resolvers.ts
@@ -1,0 +1,13 @@
+const rootQuery = {
+	hello: "Hello world",
+	message: "server running!",
+	status: 200,
+};
+
+
+export const resolvers = {
+	Query: {
+		rootQuery: () => rootQuery,
+	},
+	// Mutations: {...}
+};
