+++ server/src/graphql/typeDefs.ts
@@ -1,0 +1,13 @@
+import { gql } from "apollo-server-express";
+
+export const typeDefs = gql`
+	type RootQuery {
+		hello: String
+		message: String
+		status: Int
+	}
+
+	type Query {
+		rootQuery: RootQuery
+	}
+`;
