+++ server/src/server.ts
@@ -2,7 +2,7 @@ import { createServer } from "http";
 import express from "express";
 import { ApolloServer } from "apollo-server-express";
 
-import { typeDefs, resolvers } from "./graphql/examples/queries";
+import { typeDefs, resolvers } from "./graphql/index";
 
 const PORT = 4000;
 const app = express();
